requirements.txt
discord==2.3.2
discord.py==2.3.2
python-dotenv==1.0.1
multidict==6.0.4
aiohttp==3.9.3
aiosignal==1.3.1
frozenlist==1.4.1
yarl==1.9.4
requests==2.31.0
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.0
certifi==2023.11.17
nest_asyncio==1.6.0
yt-dlp==2023.12.30
ffmpeg==1.4